+++ src/app/pages/components/display-text-head/display-text-head.component.ts
